requirements.txt
torch>=2.0.0
torchvision>=0.15.0
paddlepaddle>=2.5.0
paddlevision>=1.0.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
PyYAML>=6.0
tensorboard>=2.13.0
matplotlib>=3.7.0
scipy>=1.10.0
gradio>=4.0.0
lpips>=0.1.4
clean-fid>=0.1.35
loguru>=0.7.0
pytest>=7.4.0 